chore(feedback-container): break down of feedback-container component
