+++ packages/chat/__stories__/feedback-container.stories.js
@@ -26,8 +26,8 @@ export default {
 export const Default = {
   args: {
     apiKey: '',
-    user: '',
-    model: '',
+    userId: '',
+    'ai-model': '',
     input: 'What is Lorem Ipsum?',
     output:
       'The standard chunk of Lorem Ipsum used since the 1500s is reproduced below for those interested. Sections 1.10.32 and 1.10.33 from "de Finibus Bonorum et Malorum" by Cicero are also reproduced in their exact original form, accompanied by English versions from the 1914 translation by H. Rackham.',
@@ -48,8 +48,8 @@ export const Default = {
       <br />
       <c4ai-feedback-container
         api-key=${args.apiKey}
-        user=${args.user}
-        ai-model=${args.model}
+        user=${args.userId}
+        ai-model=${args['ai-model']}
         input=${args.input}
         output=${args.output}>
         ${args.output}
