+++ packages/chat/components/feedback/src/feedback-container.template.ts
@@ -7,11 +7,8 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import { LitElement, html, PropertyValueMap } from 'lit';
-import { property, state } from 'lit/decorators.js';
-import { v4 as uuidv4 } from 'uuid';
-// @ts-ignore
-import styles from './feedback-container.scss?inline';
+import { html } from 'lit';
+import Flag24 from '@carbon/web-components/es/icons/flag/24';
 import '@carbon/web-components/es/components/modal/index.js';
 import '@carbon/web-components/es/components/textarea/index.js';
 import '@carbon/web-components/es/components/radio-button/index.js';
@@ -19,305 +16,111 @@ import '@carbon/web-components/es/components/button/index.js';
 import '@carbon/web-components/es/components/checkbox/index.js';
 import '@carbon/web-components/es/components/form-group/index.js';
 
-import Flag24 from '@carbon/web-components/es/icons/flag/24';
-import { FeedbackAPI } from '../../../services/feedback';
+/**
+ * Lit template for Feedback Flag
+ *
+ * @param {object} customElementClass Class functionality for the custom element
+ * @returns {TemplateResult<1>} Lit html template
+ */
+function openFeedbackFlag(customElementClass) {
+  const range = window.getSelection()?.getRangeAt(0).getBoundingClientRect();
+  const { _toggle: toggle, isModelOpen } = customElementClass;
+  return html`
+    <div
+      @click=${toggle}
+      class="feedback-flag"
+      style="top: ${-(
+        range?.height! + 8
+      )}px; left: ${range?.left}px; font-weight: bold; display: ${isModelOpen
+        ? 'none'
+        : 'block'}">
+      ${Flag24({ slot: 'icon' })}
+    </div>
+  `;
+}
 
 /**
- * Feedback Container component to record and give feedback on AI generated content
+ * Lit template for Feedback Modal
+ *
+ * @param {object} customElementClass Class functionality for the custom element
+ * @returns {TemplateResult<1>} Lit html template
  */
-export class FeedbackContainer extends LitElement {
-  static styles = styles;
+function openModal(customElementClass) {
+  let {
+    isModelOpen,
+    formData,
+    _handleTextInput: handleTextInput,
+    _handleFeedback: handleFeedback,
+    _handleTextArea: handleTextArea,
+    _handleFormData: handleFormData,
+  } = customElementClass;
 
-  /**
-   * API Key of registered application for using this component
-   */
-  @property({ attribute: 'api-key' })
-  private _api_key: string = '';
+  return html`
+    <cds-modal
+      id="feedback-modal"
+      ?open=${isModelOpen}
+      @cds-modal-closed=${() => (isModelOpen = false)}>
+      <cds-modal-header>
+        <cds-modal-close-button></cds-modal-close-button>
+        <cds-modal-heading>Provide Feedback</cds-modal-heading>
+      </cds-modal-header>
+      <cds-modal-body>
+        <cds-form-group>
+          <cds-text-input
+            label="Selected Text:"
+            value=${formData.selected_value}
+            @input=${handleTextInput}>
+          </cds-text-input>
+        </cds-form-group>
 
-  /**
-   * Unique username or id of the application
-   */
-  @property({ attribute: 'user' })
-  private _user_id: string = '';
+        <cds-form-group @cds-checkbox-changed=${handleFeedback}>
+          <cds-checkbox label-text="No problematic text found"></cds-checkbox>
+          <cds-checkbox
+            label-text="Contains HAP (e.g. hate, abusive language, profanity)"></cds-checkbox>
+          <cds-checkbox
+            label-text="Contains PII (e.g. SSN, VIN, personal address)"></cds-checkbox>
+          <cds-checkbox
+            label-text="Contains social bias (e.g. race, religion, social status, etc.)"></cds-checkbox>
+          <cds-checkbox label-text="Isn’t truthful/honest"></cds-checkbox>
+          <cds-checkbox
+            label-text="Contains Taboo Topics (eg. religion, politics etc.)"></cds-checkbox>
+          <cds-checkbox
+            label-text="Other problem (please provide detail)"></cds-checkbox>
+        </cds-form-group>
 
-  /**
-   * Model ID or Model Name for which feedback is recording
-   */
-  @property({ attribute: 'ai-model' })
-  private _model_id: string = '';
+        <cds-form-group>
+          <cds-textarea
+            rows="4"
+            style="padding:8px"
+            value=${formData.comment}
+            @input=${handleTextArea}>
+            <span slot="label-text">Comments:</span>
+          </cds-textarea>
+        </cds-form-group>
+      </cds-modal-body>
+      <cds-modal-footer>
+        <cds-modal-footer-button kind="secondary" data-modal-close
+          >Cancel</cds-modal-footer-button
+        >
+        <cds-modal-footer-button kind="primary" @click=${handleFormData}
+          >Save</cds-modal-footer-button
+        >
+      </cds-modal-footer>
+    </cds-modal>
+  `;
+}
 
-  /**
-   * User Input to the model
-   */
-  @property({ attribute: 'input' })
-  private _input: string = '';
-
-  /**
-   * Output generated by AI Model
-   */
-  @property({ attribute: 'output' })
-  private _output: string = '';
-
-  /**
-   * ID generated For a particular input and output
-   */
-  @state()
-  generation_id: string = '';
-
-  /**
-   * State variable for Feedback Model
-   */
-  @state()
-  private isModelOpen = false;
-
-  /**
-   * To store data of selected text
-   */
-  @state()
-  private selection;
-
-  /**
-   * Object for recording the feedback
-   */
-  @state()
-  private formData = {
-    id: '',
-    generation_id: '',
-    start_index: 0,
-    end_index: 0,
-    selected_value: '',
-    corrected_value: '',
-    feedback: '',
-    comment: '',
-  };
-
-  /**
-   * Instance of feedback API
-   */
-  private feedbackApi = FeedbackAPI.getInstance();
-
-  /**
-   * Array for storing checkbox values selected by user
-   */
-  private _feedbacks: string[] = [];
-
-  constructor() {
-    super();
-    this.generation_id = uuidv4();
-  }
-
-  connectedCallback(): void {
-    super.connectedCallback();
-    if (
-      this._api_key &&
-      this._model_id &&
-      this._user_id &&
-      (this._input || this._output)
-    ) {
-      const payload = {
-        id: this.generation_id,
-        api_key: this._api_key,
-        model_id: this._model_id,
-        user_id: this._user_id,
-        input_value: this._input,
-        output_value: this._output,
-      };
-      this.feedbackApi.recordGeneration(payload);
-    }
-  }
-
-  /**
-   * Click event handler that is attached to this component to get the selection / selected text
-   * @private
-   */
-  _handleTextSelection() {
-    let selection = window.getSelection()!;
-    this.selection = selection;
-    const selectedText = selection?.toString().trim();
-    this.setAttribute('selected', '');
-
-    if (selectedText) {
-      const minOffset = Math.min(selection.anchorOffset, selection.focusOffset);
-      const maxOffset = Math.max(selection.anchorOffset, selection.focusOffset);
-      this.formData.generation_id = this.generation_id;
-      this.formData.selected_value = selectedText;
-      this.formData.start_index = minOffset;
-      this.formData.end_index = maxOffset;
-    } else {
-      this.selection = null;
-      this.removeAttribute('selected');
-    }
-  }
-
-  /**
-   * Method for rendering feedback flag icon
-   */
-  openFeedbackFlag() {
-    const range = window.getSelection()?.getRangeAt(0).getBoundingClientRect();
-    return html`
-      <div
-        @click=${this._toggle}
-        class="feedback-flag"
-        style="top: ${-(
-          range?.height! + 8
-        )}px; left: ${range?.left}px; font-weight: bold; display: ${this
-          .isModelOpen
-          ? 'none'
-          : 'block'}">
-        ${Flag24({ slot: 'icon' })}
-      </div>
-    `;
-  }
-
-  /**
-   * Input event handler that is attached to the feedback (corrected value) form input
-   *
-   * @param {object} event Event object of the corrected value from input box
-   * @param {object} event.target input element
-   * @private
-   */
-  handleTextInput({ target }: Event) {
-    const { value } = target as HTMLInputElement;
-    this.formData.corrected_value = value;
-  }
-
-  /**
-   * Input event handler that is attached to the feedback (comments) form input
-   *
-   * @param {object} event Event object of the corrected value from input box
-   * @param {object} event.target input element
-   * @private
-   */
-  handleTextArea(event) {
-    this.formData.comment = event?.target.value;
-  }
-
-  /**
-   * Submit/Record the feedback data to backend
-   * @private
-   */
-  _handleFormData() {
-    if (!this.formData.corrected_value) {
-      this.formData.corrected_value = this.formData.selected_value;
-    }
-    this.feedbackApi.recordFeedback(this.formData);
-    this.selection = null;
-    this.formData = {
-      id: '',
-      generation_id: '',
-      start_index: 0,
-      end_index: 0,
-      selected_value: '',
-      corrected_value: '',
-      feedback: '',
-      comment: '',
-    };
-    this._feedbacks = [];
-    this.isModelOpen = false;
-  }
-
-  protected updated(
-    _changedProperties: PropertyValueMap<any> | Map<PropertyKey, unknown>
-  ): void {
-    if (this.isModelOpen) {
-      this.removeEventListener('click', this._handleTextSelection, false);
-    } else {
-      this.addEventListener('click', this._handleTextSelection);
-    }
-    // TODO: If we want to capture feedback while typing can be done here
-  }
-
-  /**
-   * Checkbox Input event handler that is attached to the feedback form input
-   *
-   * @param {object} event Event object of the corrected value from input box
-   * @param {object} event.target input element
-   * @private
-   */
-  handleFeedback(event) {
-    let feedback = event.target.labelText;
-    if (!this._feedbacks.includes(feedback)) {
-      this._feedbacks.push(feedback);
-    } else {
-      this._feedbacks = this._feedbacks.filter((item) => item != feedback);
-    }
-    this.formData.feedback = this._feedbacks.join(' | ');
-  }
-
-  /**
-   * Method for rendering Feedback Modal
-   */
-  openModal() {
-    this.selection = null;
-    return html`
-      <cds-modal
-        id="feedback-modal"
-        ?open=${this.isModelOpen}
-        @cds-modal-closed=${() => (this.isModelOpen = false)}>
-        <cds-modal-header>
-          <cds-modal-close-button></cds-modal-close-button>
-          <cds-modal-heading>Provide Feedback</cds-modal-heading>
-        </cds-modal-header>
-        <cds-modal-body>
-          <cds-form-group>
-            <cds-text-input
-              label="Selected Text:"
-              value=${this.formData.selected_value}
-              @input=${this.handleTextInput}>
-            </cds-text-input>
-          </cds-form-group>
-
-          <cds-form-group @cds-checkbox-changed=${this.handleFeedback}>
-            <cds-checkbox label-text="No problematic text found"></cds-checkbox>
-            <cds-checkbox
-              label-text="Contains HAP (e.g. hate, abusive language, profanity)"></cds-checkbox>
-            <cds-checkbox
-              label-text="Contains PII (e.g. SSN, VIN, personal address)"></cds-checkbox>
-            <cds-checkbox
-              label-text="Contains social bias (e.g. race, religion, social status, etc.)"></cds-checkbox>
-            <cds-checkbox label-text="Isn’t truthful/honest"></cds-checkbox>
-            <cds-checkbox
-              label-text="Contains Taboo Topics (eg. religion, politics etc.)"></cds-checkbox>
-            <cds-checkbox
-              label-text="Other problem (please provide detail)"></cds-checkbox>
-          </cds-form-group>
-
-          <cds-form-group>
-            <cds-textarea
-              rows="4"
-              style="padding:8px"
-              value=${this.formData.comment}
-              @input=${this.handleTextArea}>
-              <span slot="label-text">Comments:</span>
-            </cds-textarea>
-          </cds-form-group>
-        </cds-modal-body>
-        <cds-modal-footer>
-          <cds-modal-footer-button kind="secondary" data-modal-close
-            >Cancel</cds-modal-footer-button
-          >
-          <cds-modal-footer-button kind="primary" @click=${this._handleFormData}
-            >Save</cds-modal-footer-button
-          >
-        </cds-modal-footer>
-      </cds-modal>
-    `;
-  }
-
-  /**
-   * Method for toggling the Feedback Modal
-   */
-  _toggle() {
-    this.isModelOpen = !this.isModelOpen;
-  }
-
-  handleSlotChange() {}
-
-  render() {
-    return html`
-      ${this.selection ? this.openFeedbackFlag() : null}
-      <slot @slotchange=${this.handleSlotChange}></slot>
-      ${this.isModelOpen ? this.openModal() : ''}
-    `;
-  }
+/**
+ * Lit template for Feedback Container
+ *
+ * @param {object} customElementClass Class functionality for the custom element
+ * @returns {TemplateResult<1>} Lit html template
+ */
+export function feedbackContainerTemplate(customElementClass) {
+  const { isModelOpen, selection } = customElementClass;
+  return html`
+    ${selection ? openFeedbackFlag(customElementClass) : null}
+    <slot></slot>
+    ${isModelOpen ? openModal(customElementClass) : ''}
+  `;
 }
